Fix API routes
