+++ config/routes.rb
@@ -3,5 +3,14 @@ Rails.application.routes.draw do
   get 'multiple_clear_cert', to: 'cert_reaper/hosts#multiple_clear_cert'
   post 'submit_multiple_clear_cert',
        to: 'cert_reaper/hosts#submit_multiple_clear_cert'
-  delete 'certs', to: 'cert_reaper/api/v2/certs#destroy'
+
+  namespace :cert_reaper do
+    namespace :api do
+      scope "(:apiv)", :module => :v2, :defaults => {:apiv => 'v2'},
+          :apiv => /v1|v2/, :constraints => ApiConstraints.new(:version => 2) do
+
+        resource :certs, :only => [:destroy]
+      end
+    end
+  end
 end
